+++ model/hy_asset_node.gd
@@ -290,16 +290,26 @@ func serialize_me(schema: AssetNodesSchema, gn_lookup: Dictionary[String, GraphN
         var serialized_type_key: Variant = schema.node_types.find_key(an_type)
         if serialized_type_key and serialized_type_key.split("|", false).size() > 1:
             serialized_data["Type"] = serialized_type_key.split("|")[1]
+
         for setting_key in node_schema.get("settings", {}).keys():
-            if node_schema["settings"][setting_key]["gd_type"] == TYPE_STRING:
+            var gd_type: int = node_schema["settings"][setting_key]["gd_type"]
+            if gd_type == TYPE_STRING:
                 if not settings[setting_key]:
                     continue
-            elif node_schema["settings"][setting_key]["gd_type"] == TYPE_INT:
+            elif gd_type == TYPE_INT:
                 if settings.has(setting_key):
                     if typeof(settings[setting_key]) == TYPE_FLOAT:
                         settings[setting_key] = roundi(settings[setting_key])
                     elif typeof(settings[setting_key]) == TYPE_STRING:
                         settings[setting_key] = roundi(float(settings[setting_key]))
+            elif gd_type == TYPE_ARRAY:
+                var array_gd_type: int = node_schema["settings"][setting_key]["array_gd_type"]
+                if array_gd_type == TYPE_INT:
+                    for i in settings[setting_key].size():
+                        if typeof(settings[setting_key][i]) == TYPE_FLOAT:
+                            settings[setting_key][i] = roundi(settings[setting_key][i])
+                        elif typeof(settings[setting_key][i]) == TYPE_STRING:
+                            settings[setting_key][i] = roundi(float(settings[setting_key][i]))
             serialized_data[setting_key] = settings[setting_key]
         for conn_name in node_schema.get("connections", {}).keys():
             var num_connected: = num_connected_asset_nodes(conn_name)